BaseEntity w/ EntityDefinition interfaces // new migrations

+++ src/Todos.Application/Interfaces/IAsyncCrudRepository.cs
@@ -6,7 +6,7 @@ namespace Todos.Application.Interfaces
     public interface IAsyncCrudRepository<TEntity>
     {
         Task<IList<TEntity>> GetAllAsync();
-        Task<TEntity> GetAsync(int id);
+        Task<TEntity> GetAsync(long id);
         Task<TEntity> CreateAsync(TEntity entity);
         Task<TEntity> UpdateAsync(TEntity entity);
         Task DeleteAsync(TEntity entity);
